pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Moraxyc <me@morax.icu>"]
license = "GPL2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.12.0"
python-frontmatter = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate = "app.main:main"
poetry = "poetry.console.application:main"
